Change all uses of sets and lists of bytes to fastutils equivalents
Also changed all usages of List<Byte> to ByteList, and Set<Byte> for
ByteSet. Changed the concrete classes from ArrayList to ByteArrayList
and from LinkedHashSet to ByteOpenLinkedHashSet.

Fastutils is a library of primitive collection implementations that
are compatible with Java Collections. For motivation on the choice of
fastutils for primitive collections (as opposed to, for example,
Trove), see [this
comparison](http://java-performance.info/hashmap-overview-jdk-fastutil-goldman-sachs-hppc-koloboke-trove-january-2015/).