[ByteCompactLinkedSet] Add a linked version of byte compact set
This is a "linked" version of the ByteCompactSet. It maintains an
array of the values in the set, and uses the bitfields for fast
checking if values are present.

The main difference is that the set maintains insertion order, and
that iteration is O(n) where n is the number of elements in the set
insted of O(256).