[ExperimentResult] Change experiment output to space-separated values for cleaner output file.
Output file will no longer be a broken MAT-file. Instead it will be a space-separated values (ssv) file.
Previously we always needed to edit the output file manually by removing the last ';' character and
adding a ']' character at the end to make it a valid MAT-file. Now it will be a valid ssv-file that can be
read straight into gnuplot without modifications, or loaded into Matlab or Octave by using the

result_matrix = dlmread('filename.dat', ' ', 19, 0)

command, where 19 is the number of comment rows in the output file.