[TinyGreedyAlgorithm] Avoid the boolean list in isSingleTileHole
The list of boolean in isSingleTileHole is not actually
needed. Replace with a single boolean that is set to false when an
adjacent position that is free and inside the harmony bounds is
detected.

This improved the sum for all strategies in
TimingTest.timeAllStrategies from around 10s to around 9s.