Fix some issues with the plot scripts (plotting to the figures/ directory).
Writing figures to an output directory other than "figures/" caused the .tex and .eps files to
contain references to a directory structire that will not be present when the figures are copied
to the article project. So the output directory of the gnuplot scripts were changed back to
"figures/" and instead we move the entire figures directory to the target directory in the
process_results.sh script once all plots have been produced.

Also added the 10s run for MCE-FG/R and MCE-TR/R in experiment 4.