Add more AI players for UCB constant experiments.
We add some UCT-FG players for examining the UCB constant. The reason
for this is that the exploration term in UCB probably does not have that much
impact on UCT-TR since it can do so many playouts each move. The explorer
term in UCB probably has a higher impact on players with expensive playout
policies (e.g., UCT-FG) which have to select move based on far fewer playouts.